progress (#539)
* progress

* some renaming

* prepare grid for infinite loading

* tuning

* progress

* improve: only import from backend for hono clients

* prepare pagination queries by moving defaults to the hook, so we replace api.ts files entirely with generated sdks?

* more

* missed a piece

* experiment with sdk for getOrganizations

* progress

* fixes

* use a vite plugin to rewrite zod.gen.ts import

* ignore some ts errors

* add: global clientConfig

* remove: organizations api.ts

* remove: users api.ts

* remove: success from schema

* remove: requests api.ts

* remove: attachment api.ts

* remove: system api.ts

* remove: entities api.ts

* remove requests api.ts

* progress me/api.ts

* revert tanstack router until route issue is fixed

* remove: membership api.ts

* chore: improve queryFn readability

* remove: aut api.ts

* improve: boime & partly code by linter

* linter fixes

* dont import backend anymore

* chore: refactor clientConfig

---------

+++ backend/src/modules/requests/handlers.ts
@@ -1,10 +1,10 @@
 import { OpenAPIHono } from '@hono/zod-openapi';
-import { type SQL, and, count, eq, getTableColumns, ilike, inArray, sql } from 'drizzle-orm';
+import { and, count, eq, getTableColumns, ilike, inArray, type SQL, sql } from 'drizzle-orm';
 
 import { db } from '#/db/db';
 import { type RequestModel, requestsTable } from '#/db/schema/requests';
 import type { Env } from '#/lib/context';
-import { errorResponse } from '#/lib/errors';
+import { type ErrorType, errorResponse } from '#/lib/errors';
 import { sendSlackMessage } from '#/lib/notifications';
 import requestRoutes from '#/modules/requests/routes';
 import { getUserBy } from '#/modules/users/helpers/get-user-by';
@@ -56,7 +56,7 @@ const requestRouteHandlers = app
       wasInvited: false,
     };
 
-    return ctx.json({ success: true, data }, 200);
+    return ctx.json(data, 200);
   })
   /*
    *  Get list of requests for system admins
@@ -89,7 +89,7 @@ const requestRouteHandlers = app
 
     const items = await db.select().from(requestsQuery.as('requests')).orderBy(orderColumn).limit(Number(limit)).offset(Number(offset));
 
-    return ctx.json({ success: true, data: { items, total } }, 200);
+    return ctx.json({ items, total }, 200);
   })
   /*
    *  Delete requests
@@ -99,11 +99,15 @@ const requestRouteHandlers = app
 
     // Convert the ids to an array
     const toDeleteIds = Array.isArray(ids) ? ids : [ids];
+    if (!toDeleteIds.length) return errorResponse(ctx, 400, 'invalid_request', 'error', 'organization');
 
     // Delete the requests
     await db.delete(requestsTable).where(inArray(requestsTable.id, toDeleteIds));
 
-    return ctx.json({ success: true }, 200);
+    // Map errors of requests user is not allowed to delete
+    const errors: ErrorType[] = [];
+
+    return ctx.json({ success: true, errors }, 200);
   });
 
 export default requestRouteHandlers;
